Refactored LIMS API interface
Done to ease implementation of other LIMS API that use different backends
besides CouchDB.
